src/old/pi1/requirements.txt
AWSIoTPythonSDK 
smbus2
paho-mqtt
sparkfun-qwiic
pyyaml
boto3
m2r2